Refactor: outside user is responsible for setting lexer parsing mode
The previous design was problematic because only the parser really knows when a breakout starts/ends. Now the parser can use this information to switch parsing modes.

A consequence is that we no longer just yield return (because the parser will need to peek tokens). The parser can now peek and consume tokens to advance.